Added `name` to Interval
First step for #108

Added new `name` arg to `Interval` constructor, and will set `self.name` to that, if it is passed.

Still need to think about where in the repr this attribute goes, and it will need to be caught in various methods that call `Interval`, such as `Striplog._build_list_of_intervals`. There do not seem to be many, on a quick search.